Refactored project names to snake case
