+++ osirix/exceptions.py
