+++ osirix/osirix_utils.py
@@ -3,10 +3,10 @@ from typing import List, Tuple
 
 import grpc
 import sys
-from osirix.Exceptions import GrpcException
-from osirix.ViewerController import ViewerController, VRController
-from osirix.BrowserController import BrowserController
-from osirix.ResponseProcessor import ResponseProcessor
+from osirix.exceptions import GrpcException
+from osirix.viewer_controller import ViewerController, VRController
+from osirix.browser_controller import BrowserController
+from osirix.response_processor import ResponseProcessor
 
 # sys.path.append("./pb2/")
